+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to OpenEMC will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.0
+### Added
+- pwm: new `openemc,complementary-invert` device-tree property
+  to configure per channel whether complementary output (if present) 
+  is inverted versus main output
+
 ## 1.2.0 - 2026-05-12
 ### Added
 - pwm: per-channel duty cycle ramp time, configurable via the new